requirements.txt
openai
flask
flask-script
gunicorn
requests
python-dotenv
textblob
